feat(experimental-utils): expose our RuleTester extension
There are a few things we've built in (like automatically adding a filename) that others might find value in.
Plus I want to use it in the internal plugin.
