+++ packages/eslint-plugin-internal/tests/RuleTester.ts
@@ -1,22 +1,10 @@
-import { TSESLint, ESLintUtils } from '@typescript-eslint/experimental-utils';
+import { ESLintUtils } from '@typescript-eslint/experimental-utils';
+import path from 'path';
 
-const { batchedSingleLineTests } = ESLintUtils;
-
-const parser = '@typescript-eslint/parser';
-
-type RuleTesterConfig = Omit<TSESLint.RuleTesterConfig, 'parser'> & {
-  parser: typeof parser;
-};
-class RuleTester extends TSESLint.RuleTester {
-  // as of eslint 6 you have to provide an absolute path to the parser
-  // but that's not as clean to type, this saves us trying to manually enforce
-  // that contributors require.resolve everything
-  constructor(options: RuleTesterConfig) {
-    super({
-      ...options,
-      parser: require.resolve(options.parser),
-    });
-  }
+function getFixturesRootDir(): string {
+  return path.join(__dirname, 'fixtures');
 }
 
-export { RuleTester, batchedSingleLineTests };
+const { batchedSingleLineTests, RuleTester } = ESLintUtils;
+
+export { RuleTester, batchedSingleLineTests, getFixturesRootDir };
