+++ packages/experimental-utils/src/eslint-utils/index.ts
@@ -2,4 +2,5 @@ export * from './applyDefault';
 export * from './batchedSingleLineTests';
 export * from './getParserServices';
 export * from './RuleCreator';
+export * from './RuleTester';
 export * from './deepMerge';
